legacy/requirements.txt
certifi
chardet
click
Flask==2.2.2
Flask-SQLAlchemy >= 2.5.1
Flask-WTF==1.0.1
gunicorn==19.7.1
idna==2.5
itsdangerous==2.0.1
Jinja2
MarkupSafe
olefile==0.44
qrcode
Pillow==9.2.0
requests
# requests==2.20.0
# SQLAlchemy >= 2.5.1
## urllib3==1.24.2
urllib3
Werkzeug
WTForms==3.0.1
python-dateutil
python-dotenv
loguru